refactor: 컴포넌트 파일 구조 정리 및 import 경로 수정
