+++ src/App.jsx
@@ -8,7 +8,7 @@ import { useState } from "react";
 import Home from "./pages/Home/Home.jsx";
 import Prompt from "./pages/Prompt/Prompt.jsx";
 import SignUp from "./pages/SignUp/SignUp.jsx";
-import Header from "./components/Header.jsx";
+import Header from "./components/Header/Header.jsx";
 import ProfilePage from "./pages/ProfilePage/ProfilePage.jsx";
 import ReservationsPage from "./pages/ReservationsPage/ReservationsPage.jsx";
 import HistoryPage from "./pages/HistoryPage/HistoryPage.jsx";
